# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Gluco Guardian API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /doctor/auth/register:
    post:
      operationId: doctorAuthRegister
      tags: [doctor]
      summary: Register a doctor account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DoctorRegisterRequest"
      responses:
        "201":
          description: Doctor registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DoctorRegisterResponse"
        "409":
          description: Email already registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /doctor/auth/login:
    post:
      operationId: doctorAuthLogin
      tags: [doctor]
      summary: Doctor account login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DoctorAccountLoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DoctorAccountLoginResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /doctor/auth/logout:
    post:
      operationId: doctorAuthLogout
      tags: [doctor]
      summary: Revoke the current doctor session
      security:
        - doctorBearerAuth: []
      responses:
        "200":
          description: Logged out
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /doctor/me:
    get:
      operationId: getDoctorMe
      tags: [doctor]
      summary: Get the authenticated doctor profile
      security:
        - doctorBearerAuth: []
      responses:
        "200":
          description: Doctor profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DoctorProfile"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /doctor/me/patients:
    get:
      operationId: listDoctorLinkedPatients
      tags: [doctor]
      summary: List patients linked to the authenticated doctor
      security:
        - doctorBearerAuth: []
      responses:
        "200":
          description: Linked patients
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DoctorLinkedPatientsResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /doctor/me/patients/link:
    post:
      operationId: linkDoctorPatient
      tags: [doctor]
      summary: Link a patient using their access code
      security:
        - doctorBearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DoctorLinkPatientRequest"
      responses:
        "201":
          description: Patient linked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DoctorLinkPatientResponse"
        "200":
          description: Patient was already linked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DoctorLinkPatientResponse"
        "404":
          description: Invalid access code
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /doctor/me/patients/{accessCode}:
    delete:
      operationId: unlinkDoctorPatient
      tags: [doctor]
      summary: Revoke a linked patient
      security:
        - doctorBearerAuth: []
      parameters:
        - in: path
          name: accessCode
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Link revoked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "404":
          description: Link not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /doctor/login:
    post:
      operationId: doctorLogin
      tags: [doctor]
      deprecated: true
      summary: "[Deprecated] Legacy access-code check — use doctor auth + link flow"
      description: >
        Does not create a doctor session and no longer grants access to patient
        snapshot routes. Use POST /doctor/auth/login and POST /doctor/me/patients/link.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DoctorLoginRequest"
      responses:
        "200":
          description: Legacy response (deprecated)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DoctorLoginResponse"
        "401":
          description: Invalid access code
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /doctor/sync:
    post:
      operationId: syncPatientData
      tags: [doctor]
      summary: Mobile app syncs patient snapshot to server
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PatientSnapshot"
      responses:
        "200":
          description: Sync successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /doctor/patient/{accessCode}:
    get:
      operationId: getPatientData
      tags: [doctor]
      summary: Get patient snapshot for a linked access code
      security:
        - doctorBearerAuth: []
      parameters:
        - in: path
          name: accessCode
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Patient data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PatientSnapshot"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Doctor is not linked to this patient
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: No patient data found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /doctor/messages/{accessCode}:
    get:
      operationId: getDoctorMessages
      tags: [doctor]
      summary: Get messages for a linked patient
      security:
        - doctorBearerAuth: []
      parameters:
        - in: path
          name: accessCode
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Messages
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DoctorMessagesResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Doctor is not linked to this patient
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: sendDoctorMessage
      tags: [doctor]
      summary: Doctor sends a message to a linked patient
      security:
        - doctorBearerAuth: []
      parameters:
        - in: path
          name: accessCode
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendMessageRequest"
      responses:
        "200":
          description: Message sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DoctorMessage"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Doctor is not linked to this patient
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  securitySchemes:
    doctorBearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
      description: Bearer token from POST /doctor/auth/login
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
        - message
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    DoctorLoginRequest:
      type: object
      properties:
        accessCode:
          type: string
      required:
        - accessCode
    DoctorLoginResponse:
      type: object
      properties:
        success:
          type: boolean
        accessCode:
          type: string
        patientName:
          type: string
        hasData:
          type: boolean
        deprecated:
          type: boolean
        migration:
          type: string
      required:
        - success
        - accessCode
        - hasData
    DoctorRegisterRequest:
      type: object
      properties:
        email:
          type: string
        passwordHash:
          type: string
        displayName:
          type: string
        institution:
          type: string
      required:
        - email
        - passwordHash
        - displayName
    DoctorRegisterResponse:
      type: object
      properties:
        doctorId:
          type: string
      required:
        - doctorId
    DoctorAccountLoginRequest:
      type: object
      properties:
        email:
          type: string
        passwordHash:
          type: string
      required:
        - email
        - passwordHash
    DoctorProfile:
      type: object
      properties:
        doctorId:
          type: string
        email:
          type: string
        displayName:
          type: string
        institution:
          type: string
      required:
        - doctorId
        - email
        - displayName
    DoctorAccountLoginResponse:
      type: object
      properties:
        token:
          type: string
        expiresAt:
          type: number
        doctor:
          $ref: "#/components/schemas/DoctorProfile"
      required:
        - token
        - expiresAt
        - doctor
...
    DoctorLinkedPatientsResponse:
      type: object
      properties:
        patients:
          type: array
          items:
            $ref: "#/components/schemas/DoctorLinkedPatient"
      required:
        - patients
    DoctorLinkPatientRequest:
      type: object
      properties:
        accessCode:
          type: string
      required:
        - accessCode
    DoctorLinkPatientResponse:
      type: object
      properties:
        accessCode:
          type: string
        displayName:
          type: string
        patientUserId:
          type: string
        linkedAt:
          type: number
        alreadyLinked:
          type: boolean
        hasData:
          type: boolean
        syncedAt:
          type: string
      required:
        - accessCode
        - linkedAt
        - alreadyLinked
        - hasData
...
    DoctorMessagesResponse:
      type: object
      properties:
        messages:
          type: array
          items:
            $ref: "#/components/schemas/DoctorMessage"
      required:
        - messages
...
    PatientSnapshot:
      type: object
      properties:
        accessCode:
          type: string
        profile:
          $ref: "#/components/schemas/PatientProfile"
        glucoseReadings:
          type: array
          items:
            $ref: "#/components/schemas/CGMReading"
        insulinLog:
          type: array
          items:
            $ref: "#/components/schemas/InsulinLogEntry"
        foodLog:
          type: array
          items:
            $ref: "#/components/schemas/FoodLogEntry"
        messages:
          type: array
          items:
            $ref: "#/components/schemas/DoctorMessage"
        alertPreferences:
          type: object
          properties:
            lowThreshold:
              type: number
            highThreshold:
              type: number
            urgentLowThreshold:
              type: number
            urgentHighThreshold:
              type: number
        syncedAt:
          type: string
      required:
        - accessCode
        - profile
        - glucoseReadings
        - insulinLog
        - foodLog
        - messages
        - syncedAt